\name{candoo-package}
\alias{candoo-package}
\alias{candoo}
\docType{package}
\title{
Candoo is the personal package for duncandoo
}
\description{
Personal package containing functions wot I wrote and other stuff. Notably my attempt at a rmarkdown render format for reveal.js slides
}
\details{
\tabular{ll}{
Package: \tab candoo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-30\cr
License: \tab GNU Affero GPL v3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Duncan McPherson

Maintainer: Duncan <duncan@mcpherson>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
