% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/reveal.js_format.R
\name{reveal.js_format}
\alias{reveal.js_format}
\title{An rmarkdown format for producing reveal.js slides}
\usage{
reveal.js_format(toc = FALSE)
}
\arguments{
\item{toc}{Do you want a Table of Contents? Defaults to FALSE.}
}
\description{
This function allows you to render rmarkdown into reveal.js slide format.
}
\examples{
reveal.js_format()
Put a YAML block at the top of your .rmd file
---
title: "My Presentation"
output:
  Rcandoo::reveal.js_format:
    toc: false
---

render('My Presentation.Rmd')
}
\keyword{reveal.js}

