% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/revealjs_format.R
\name{revealjs_format}
\alias{revealjs_format}
\title{An rmarkdown format for producing reveal.js slides}
\usage{
revealjs_format(keep_md = FALSE, theme = "moon")
}
\arguments{
\item{keep_md}{Do you want to keep the .md intermediate file? Defaults to FALSE.}

\item{theme}{Which reveal.js theme do you want: options are Black, White, League, Sky, Beige, Simple, Serif, Night, Moon, Solarized}
}
\description{
This function allows you to render rmarkdown into reveal.js slide format.
}
\examples{
revealjs_format()
Put a YAML block at the top of your .rmd file
---
title: "My Presentation"
output:
  Rcandoo::revealjs_format:
    theme: moon
---

render('My Presentation.Rmd')
}
\keyword{reveal.js}

